ready : 재편 준비 (기존 로그인 로직 passport화 )
