+++ src/server/middleware/passport/index.ts
@@ -24,6 +24,7 @@ export default () => {
     passport.deserializeUser((id, done) => {
         // req.session에 저장된 사용자 아이디를 바탕으로 DB 조회로 사용자 정보를 얻어낸 후 req.user에 저장. 
         // 즉, id를 sql로 조회해서 전체 정보를 가져오는 복구 로직이다.
+
         /*User.findOne({ where: { id } })
         .then(user => done(null, user)) //? done()이 되면 이제 다시 req.login(user, ...) 쪽으로 되돌아가 다음 미들웨어를 실행하게 된다.
         .catch(err => done(err));*/
